+++ src/features/articles/components/ArticleItem.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable tailwindcss/no-custom-classname */
 import {
   ActionIcon,
   Anchor,
@@ -5,7 +6,6 @@ import {
   createStyles,
   Group,
   Image,
-  Stack,
   Text,
 } from '@mantine/core';
 import { IconBookmark, IconHeart, IconShare } from '@tabler/icons';
@@ -16,6 +16,12 @@ import { useNavigate } from 'react-router-dom';
 import { ReactComponent as Zenn } from '/src/assets/zenn.svg';
 
 const useStyles = createStyles((theme) => ({
+  category: {
+    fontSize: theme.fontSizes.xs,
+    [theme.fn.smallerThan('xs')]: {
+      fontSize: '10px',
+    },
+  },
   card: {
     backgroundColor:
       theme.colorScheme === 'dark' ? theme.colors.dark[7] : theme.white,
@@ -32,39 +38,34 @@ const useStyles = createStyles((theme) => ({
   },
 }));
 
-const ArticleDetail = ({ image, categories, title, date, media }: Article) => {
+const ArticleItem = ({ image, categories, title, date, media }: Article) => {
   const { classes, theme } = useStyles();
   const navigate = useNavigate();
   return (
     <Card
       radius="md"
-      className="hover:bg-m_gray-1 bg-m_gray-0 max-w-md py-0 hover:cursor-pointer"
+      className="bg-m_gray-0 hover:bg-m_gray-1 max-w-md py-0 hover:cursor-pointer"
       // component="a"
       // href="https://zenn.dev/"
       // target="_blank"
     >
-      <Stack>
-        <div className="pt-4">
-          <a href="https://zenn.dev/" target="_blank" rel="noreferrer">
-            <Image src={image} />
-          </a>
-        </div>
+      <Group noWrap spacing={0} className="">
         <div className="">
-          <div className="mb-2 flex space-x-2">
+          <div className="xs:space-x-2 mb-2 flex space-x-1">
             {categories.map((category) => (
               <Anchor
                 key={category.title}
                 color="dimmed"
                 weight={700}
-                size="xs"
-                onClick={() => navigate(`/categories/${category.path}`)}
+                className={classes.category}
+                onClick={() => navigate(`/categories/${category.path}/all`)}
               >
                 #{category.title}
               </Anchor>
             ))}
           </div>
           <Anchor
-            className="font-bold leading-tight text-black"
+            className="max-xs:text-sm font-bold leading-tight text-black"
             mt="xs"
             mb="md"
             href="https://zenn.dev/"
@@ -72,40 +73,41 @@ const ArticleDetail = ({ image, categories, title, date, media }: Article) => {
           >
             {title}
           </Anchor>
-          <Group noWrap spacing="xs" className="mt-2 justify-between">
+          <Group noWrap spacing="xs">
             <Group spacing="xs" noWrap>
               <ActionIcon size="md">
                 <Zenn style={{ color: '#3EA8FF' }} />
               </ActionIcon>
               <Text size="xs">{media}</Text>
-              <Text size="xs" color="dimmed">
-                {date}
-              </Text>
             </Group>
-            <Group className="">
-              <ActionIcon>
-                <IconHeart size={18} color={theme.colors.red[6]} stroke={1.5} />
-              </ActionIcon>
-              <ActionIcon>
-                <IconBookmark
-                  size={18}
-                  color={theme.colors.yellow[6]}
-                  stroke={1.5}
-                />
-              </ActionIcon>
-              <ActionIcon>
-                <IconShare
-                  size={16}
-                  color={theme.colors.blue[6]}
-                  stroke={1.5}
-                />
-              </ActionIcon>
-            </Group>
+            <Text size="xs" color="dimmed">
+              {date}
+            </Text>
           </Group>
         </div>
-      </Stack>
+        <div className="pt-4">
+          <a href="https://zenn.dev/" target="_blank" rel="noreferrer">
+            <Image src={image} height={100} width={140} />
+          </a>
+          <Group className="justify-between px-2">
+            <ActionIcon>
+              <IconHeart size={18} color={theme.colors.red[6]} stroke={1.5} />
+            </ActionIcon>
+            <ActionIcon>
+              <IconBookmark
+                size={18}
+                color={theme.colors.yellow[6]}
+                stroke={1.5}
+              />
+            </ActionIcon>
+            <ActionIcon>
+              <IconShare size={16} color={theme.colors.blue[6]} stroke={1.5} />
+            </ActionIcon>
+          </Group>
+        </div>
+      </Group>
     </Card>
   );
 };
 
-export default ArticleDetail;
+export default ArticleItem;
